+++ dotnet/src/Logging/Logging/SmartStandards/LoggingHelper.cs
@@ -9,10 +9,40 @@ namespace Logging.SmartStandards {
       return StatusToFormattedLogEntry(4, 0, statusMessage, null);
     }
 
+    /// <summary>
+    ///   Renders a "Formatted Log Entry" (see conventions) from a statusMessageTemplate + statusMessageArgs combo.
+    ///   The syntax is: "[LevelAsAlpha3]Id:ResolvedMessage"
+    /// </summary>
+    /// <param name="level">
+    ///   The numeric level of the status message. Will be rendered as Alpha3 representation:
+    ///   0: "Trc"
+    ///   1: "Dbg"
+    ///   2: "Inf"
+    ///   3: "Wrn"
+    ///   4: "Err"
+    ///   5 : "Cri" (will be used as fallback for all other values)
+    /// </param>
+    /// <param name="returnCode"> The passed return code will be used as id. </param>
+    /// <param name="statusMessageTemplate"> 
+    ///   The statusMessageTemplate, the placeholders of which will be resolved.
+    ///   null is equivalent to passing an empty string (the statusMessageArgs are ignored).
+    /// </param>
+    /// <param name="statusMessageArgs">
+    ///   The values that are used for placeholder resolving.
+    ///   null is equivalent to passing an empty array.
+    /// </param>
+    /// <returns> A string like "[Inf]4711:Hello world, the answer is 42.". </returns>
     public static string StatusToFormattedLogEntry(int level, int returnCode, string statusMessageTemplate, object[] statusMessageArgs) {
 
-      StringBuilder sb = new StringBuilder((int)(statusMessageTemplate.Length * 1.5));
+      int assumedLength;
 
+      if (statusMessageTemplate != null) {
+        assumedLength = (int)(statusMessageTemplate.Length * 1.5);
+      } else {
+        assumedLength = 10;
+      }
+
+      StringBuilder sb = new StringBuilder(assumedLength);
 
       // "[ERR]4711:File not found on Disk."